+++ examples/document-search/basic.py
@@ -33,7 +33,7 @@ To run the script, execute the following command:
 import asyncio
 
 from ragbits.core import audit
-from ragbits.core.embeddings.litellm import LiteLLMEmbedder
+from ragbits.core.embeddings.dense import LiteLLMEmbedder
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search import DocumentSearch
 from ragbits.document_search.documents.document import DocumentMeta