+++ examples/evaluation/document-search/basic/evaluate.py
@@ -38,7 +38,7 @@ config = {
                         "score_threshold": -1.2,
                     },
                     "embedder": {
-                        "type": "ragbits.core.embeddings.litellm:LiteLLMEmbedder",
+                        "type": "ragbits.core.embeddings.dense:LiteLLMEmbedder",
                         "config": {
                             "model": "text-embedding-3-small",
                         },